Eps files changes

See merge request mpaluch2/zpr_eps!6

+++ src/Scalar2DRepresentation.cc
@@ -22,25 +22,25 @@ std::string Resolution::toString() const {
     return s.str();
 }
 
-void Point::setX( int new_x ) {
+void Point::setX( double new_x ) {
     //@todo maybe exception?
     if (new_x >= 0)
         Scalar2D::setX(new_x);
 }
 
-void Point::setY( int new_y ) {
+void Point::setY( double new_y ) {
     //@todo maybe exception?
     if (new_y >= 0)
         Scalar2D::setY(new_y);
 }
 
-void Resolution::setX( int new_x ) {
+void Resolution::setX( double new_x ) {
     //@todo maybe exception?
     if (new_x >= 0)
         Scalar2D::setX(new_x);
 }
 
-void Resolution::setY( int new_y ) {
+void Resolution::setY( double new_y ) {
     //@todo maybe exception?
     if (new_y >= 0)
         Scalar2D::setY(new_y);
